pyproject.toml
[project]
name = "fluxconvert"
version = "0.1.0"
description = "Media conversion orchestrator with CLI, GUI, and local API"
readme = "readme.txt"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "camerensmith" }
]
dependencies = [
  "typer[all]>=0.12.3",
  "fastapi>=0.112.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.5.0"
]

[project.scripts]
fluxconvert = "fluxconvert.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.packages.find]
where = ["."]
include = ["fluxconvert*"]

